sap-iflow-rag-codebert/requirements.txt
supabase
sentence-transformers
transformers
cohere
pandas
numpy
scikit-learn
matplotlib
seaborn
torch
fastapi
uvicorn
python-dotenv
